pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
    "supabase>=2.13.0",
    "twilio>=9.4.6",
    "yfinance>=0.2.54",
]
